requirements.txt
alabaster==0.7.8
apturl==0.5.2
asn1crypto==0.24.0
Babel==2.4.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
click==6.7
colorama==0.3.7
command-not-found==0.3
crypto==1.4.1
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
docutils==0.14
feedparser==5.2.1
html5lib==0.999999999
httplib2==0.9.2
idna==3.3
imagesize==0.7.1
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.6.4
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
Naked==0.1.31
netifaces==0.10.4
numpy==1.19.5
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
onboard==1.4.1
pexpect==4.2.1
Pillow==8.4.0
protobuf==3.0.0
py==1.11.0
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.12.0
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
pyshark==0.4.3
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
PyX==0.15
pyxdg==0.25
PyYAML==6.0
reportlab==3.4.0
requests==2.26.0
requests-unixsocket==0.1.5
roman==2.0.0
scapy==2.4.5
scour==0.36
SecretStorage==2.3.1
shellescape==3.8.1
simple-cdd==0.6.5
simplejson==3.13.2
six==1.11.0
Sphinx==1.6.7
system-service==0.3
systemd-python==234
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.26.7
usb-creator==0.3.3
wadllib==1.3.2
webencodings==0.5
xdiagnose==3.8.8
xkit==0.0.0
zope.interface==4.3.2
